+++ src/directives/selectFormControl/selectFormControl.directive.ts
@@ -1,11 +1,13 @@
 import {Directive, effect, inject, input, InputSignal, model, ModelSignal, OnDestroy, untracked} from '@angular/core';
 import {FormValueControl} from '@angular/forms/signals';
 import {LOGGER, Logger} from '@anglr/common';
+import {nameof} from '@jscrpt/common';
 import {isEqual} from 'lodash-es';
 import {Subscription} from 'rxjs';
 
 import {Select} from '../../components/select/select.component';
 import {SelectPluginType} from '../../misc/enums';
+import {TempValue} from '../../interfaces';
 
 /**
  * Form value control for Select
@@ -27,6 +29,11 @@ export class SelectFormControl<TValue = unknown, TPublicValue = TValue> implemen
      * Subscriptions created during initialization
      */
     protected initSubscriptions: Subscription = new Subscription();
+
+    /**
+     * Temporary value used for storing value during async write
+     */
+    protected tmpValue: TempValue<TPublicValue|TPublicValue[]> = {};
 
     //######################### public properties - implementation of FormValueControl #########################
 
@@ -62,6 +69,7 @@ export class SelectFormControl<TValue = unknown, TPublicValue = TValue> implemen
 
         effect(() => this.select.selectOptions.readonly = this.disabled() || this.readonly());
 
+        //setting value
         effect(() =>
         {
             if(!this.select.initialized())
@@ -72,16 +80,19 @@ export class SelectFormControl<TValue = unknown, TPublicValue = TValue> implemen
             const value = this.value();
             this.logger.verbose('Select: Form control: control value changed "{{@(4)value}}"', {value});
 
-            untracked(() =>
+            untracked(async () =>
             {
                 if(!isEqual(this.select.getPlugin(SelectPluginType.ValueHandler).value(), value))
                 {
+                    this.tmpValue.value = value;
                     this.logger.verbose('Select: Form control: setting select value "{{@(4)value}}"', {value});
-                    this.select.getPlugin(SelectPluginType.ValueHandler).setValue(value);
+                    await this.select.getPlugin(SelectPluginType.ValueHandler).setValue(value);
+                    delete this.tmpValue.value;
                 }
             });
         });
 
+        //getting value
         effect(() =>
         {
             if(!this.select.initialized())
@@ -89,8 +100,13 @@ export class SelectFormControl<TValue = unknown, TPublicValue = TValue> implemen
                 return;
             }
 
-            const value = this.select.getPlugin(SelectPluginType.ValueHandler).value();
+            let value = this.select.getPlugin(SelectPluginType.ValueHandler).value();
             this.logger.verbose('Select: Form control: select value changed "{{@(4)value}}"', {value});
+
+            if(nameof<TempValue<TPublicValue|TPublicValue[]>>('value') in this.tmpValue)
+            {
+                value = this.tmpValue.value;
+            }
 
             untracked(() =>
             {