+++ src/directives/withDirectAccess/withDirectAccess.directive.ts
@@ -1,8 +1,10 @@
 import {Directive, effect, model, ModelSignal, untracked} from '@angular/core';
+import {nameof} from '@jscrpt/common';
 import {isEqual} from 'lodash-es';
 
 import {SelectPluginType} from '../../misc/enums';
 import {Select} from '../../components/select/select.component';
+import {TempValue} from '../../interfaces';
 
 /**
  * Directive that allows direct access to select properties using inputs, outputs
@@ -13,6 +15,13 @@ import {Select} from '../../components/select/select.component';
 })
 export class WithDirectAccess<TValue = unknown, TPublicValue = TValue>
 {
+    //######################### protected fields #########################
+
+    /**
+     * Temporary value used for storing value during async write
+     */
+    protected tmpValue: TempValue<TPublicValue|TPublicValue[]> = {};
+
     //######################### public properties - inputs/outputs #########################
 
     /**
@@ -23,6 +32,7 @@ export class WithDirectAccess<TValue = unknown, TPublicValue = TValue>
     //######################### constructor #########################
     constructor(select: Select<TValue, TPublicValue>,)
     {
+        //setting value
         effect(() =>
         {
             if(!select.initialized())
@@ -32,15 +42,18 @@ export class WithDirectAccess<TValue = unknown, TPublicValue = TValue>
 
             const value = this.value();
 
-            untracked(() =>
+            untracked(async () =>
             {
                 if(!isEqual(select.getPlugin(SelectPluginType.ValueHandler).value(), value))
                 {
-                    select.getPlugin(SelectPluginType.ValueHandler).setValue(value);
+                    this.tmpValue.value = value;
+                    await select.getPlugin(SelectPluginType.ValueHandler).setValue(value);
+                    delete this.tmpValue.value;
                 }
             });
         });
 
+        //getting value
         effect(() =>
         {
             if(!select.initialized())
@@ -48,7 +61,12 @@ export class WithDirectAccess<TValue = unknown, TPublicValue = TValue>
                 return;
             }
 
-            const value = select.getPlugin(SelectPluginType.ValueHandler).value();
+            let value = select.getPlugin(SelectPluginType.ValueHandler).value();
+
+            if(nameof<TempValue<TPublicValue|TPublicValue[]>>('value') in this.tmpValue)
+            {
+                value = this.tmpValue.value;
+            }
 
             untracked(() =>
             {