+++ src/interfaces/tempValue/tempValue.interface.ts
@@ -1,0 +1,12 @@
+/**
+ * Used for storing temporary value during async write
+ */
+export interface TempValue<TValue>
+{
+    //######################### properties #########################
+
+    /**
+     * Stored value
+     */
+    value?: TValue|undefined|null;
+}